Update calls throughout pkg for fileio => io rename
